move parse_rif_counters to a common utilities
the motivation is to avoid cross-feature dependency
