+++ tests/common/utilities.py
@@ -36,6 +36,7 @@ from tests.common.cache import FactsCache
 from tests.common.helpers.constants import UPSTREAM_NEIGHBOR_MAP, UPSTREAM_ALL_NEIGHBOR_MAP
 from tests.common.helpers.constants import DOWNSTREAM_NEIGHBOR_MAP, DOWNSTREAM_ALL_NEIGHBOR_MAP
 from tests.common.helpers.assertions import pytest_assert
+from tests.common.portstat_utilities import parse_column_positions
 from netaddr import valid_ipv6
 
 logger = logging.getLogger(__name__)
@@ -1549,3 +1550,50 @@ def cleanup_prev_images(duthost):
     current_os_version = duthost.shell('sonic_installer list | grep Current | cut -f2 -d " "')['stdout']
     duthost.shell("sonic_installer set-next-boot {}".format(current_os_version), module_ignore_errors=True)
     duthost.shell("sonic_installer cleanup -y", module_ignore_errors=True)
+
+
+def parse_rif_counters(output_lines):
+    """Parse the output of "show interfaces counters rif" command
+    Args:
+        output_lines (list): The output lines of "show interfaces counters rif" command
+    Returns:
+        list: A dictionary, key is interface name, value is a dictionary of fields/values
+    """
+
+    header_line = ''
+    separation_line = ''
+    separation_line_number = 0
+    for idx, line in enumerate(output_lines):
+        if line.find('----') >= 0:
+            header_line = output_lines[idx - 1]
+            separation_line = output_lines[idx]
+            separation_line_number = idx
+            break
+
+    try:
+        positions = parse_column_positions(separation_line)
+    except Exception:
+        logger.error('Possibly bad command output')
+        return {}
+
+    headers = []
+    for pos in positions:
+        header = header_line[pos[0]:pos[1]].strip().lower()
+        headers.append(header)
+
+    if not headers:
+        return {}
+
+    results = {}
+    for line in output_lines[separation_line_number + 1:]:
+        portstats = []
+        for pos in positions:
+            portstat = line[pos[0]:pos[1]].strip()
+            portstats.append(portstat)
+
+        intf = portstats[0]
+        results[intf] = {}
+        for idx in range(1, len(portstats)):  # Skip the first column interface name
+            results[intf][headers[idx]] = portstats[idx].replace(',', '')
+
+    return results
