+++ tests/snmp/test_snmp_interfaces.py
@@ -2,8 +2,7 @@ import logging
 import pytest
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.helpers.snmp_helpers import get_snmp_facts
-from tests.common.utilities import wait_until
-from tests.ip.ip_util import parse_rif_counters
+from tests.common.utilities import parse_rif_counters, wait_until
 
 pytestmark = [
     pytest.mark.topology('any'),
